Update20250511

+++ set-secrets.sh
@@ -3,13 +3,13 @@ set -euf -o pipefail
 # set -x
 
 
-echo "Falls du das wirklich machen willst, dann kommentier nachfolgende Zeile aus!"
-exit 1
+test -f secrets.env && (echo "Please delete secrets.env and rerun." && exit 1)
+
+set -o allexport; source .env; set +o allexport
+COMPOSE="$COMPOSE -f docker-compose.yml -f docker-compose.prod.yml"
 
 # stop and remove all containers, otherwise we can't pass the new parameters as environment variables
-docker-compose down
-
-source .env
+$COMPOSE down
 
 # clear the file
 echo >secrets.env
@@ -28,7 +28,7 @@ export MQTT_PASSWD_CONTROLLER
 declare -p MQTT_PASSWD_CONTROLLER >>secrets.env
 rm -f mosquitto/config/mosquitto.passwd
 touch mosquitto/config/mosquitto.passwd  # otherwise a directory will be created
-docker-compose run --rm mqtt mosquitto_passwd -b /mosquitto/config/mosquitto.passwd controller "$MQTT_PASSWD_CONTROLLER"
+$COMPOSE run --rm mqtt mosquitto_passwd -b /mosquitto/config/mosquitto.passwd controller "$MQTT_PASSWD_CONTROLLER"
 echo ::endgroup::
 
 
@@ -38,7 +38,7 @@ echo ::group::pgSQL Superuser Password
 POSTGRES_PASSWORD="$(generate_password)"
 export POSTGRES_PASSWORD
 declare -p POSTGRES_PASSWORD >>secrets.env
-docker-compose run --rm db /docker-postgres-run-command.sh /update_superuser.sh
+$COMPOSE run --rm db /docker-postgres-run-command.sh /update_superuser.sh
 echo ::endgroup::
 
 echo ::group::pgSQL Django Password
@@ -46,7 +46,7 @@ POSTGRES_PASSWORD_DJANGO="$(generate_password)"
 export POSTGRES_PASSWORD_DJANGO
 declare -p POSTGRES_PASSWORD_DJANGO >>secrets.env
 USER="${POSTGRES_USER_DJANGO}" PASSWORD="${POSTGRES_PASSWORD_DJANGO}" DB="${POSTGRES_DB_DJANGO}" \
-  docker-compose run --rm \
+  $COMPOSE run --rm \
   -e USER -e PASSWORD -e DB \
   db /docker-postgres-run-command.sh /update_other_user.sh
 echo ::endgroup::
